--- js/isUnique.js
@@ -1,3 +1,0 @@
-const isUnique = (array, value) => !array.includes(value);
-
-export {isUnique};
